room/create実装
