+++ create_room.sql
@@ -2,6 +2,8 @@ DROP TABLE IF EXISTS `room`;
 CREATE TABLE `room` (
   `room_id` bigint NOT NULL AUTO_INCREMENT,
   `live_id` int NOT NULL,
-  `joined_user_count` int NOT NULL,
-  PRIMARY KEY (`room_id`)
+  `owner` bigint NOT NULL,
+  `status` int NOT NULL,
+  PRIMARY KEY (`room_id`),
+  UNIQUE KEY `owner` (`owner`)
 );
